errorLogger를 공통로직으로 빼기
