+++ game-server/src/lobby/lobbyApi.ts
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { logDetailedError } from "../utils/errorLogger";
+import { logDetailedError } from "@beyondworm/shared";
 import { v4 as uuidv4 } from "uuid";
 
 const LOBBY_SERVER_URL = process.env.LOBBY_SERVER_URL || "http://localhost:3000";
@@ -8,7 +8,7 @@ const SERVER_ID = process.env.SERVER_ID || `server.${uuidv4()}`;
 export async function registerWithLobby() {
     const address = `http://localhost:${process.env.PORT || 3001}`;
     try {
-        await axios.post(`${LOBBY_SERVER_URL}/server/register`, {
+        await axios.post(`${LOBBY_SERVER_URL}/server`, {
             serverId: SERVER_ID,
             address: address,
         });
